refactor: move HttpError to shared so both callers throw one class
The error-factory argument existed only because HttpError lived under
gen/src/image/, out of reach of enter. It is a plain Error subclass with
no gen-specific imports, so it moves to shared/ and the shared download
helper throws it directly.

enter loses nothing: throwEndpointTestError already flattens any Error to
HTTPException(400, {message}), so the 502 status is simply ignored there
while gen's funnel keeps rendering it.

38 of the 41 changed files are one-line import rewrites.
