+++ gen.pollinations.ai/src/image/createAndReturnImages.ts
@@ -4,7 +4,7 @@ import {
     type ServerType,
 } from "./availableServers.ts";
 import { getImageEnv } from "./env.ts";
-import { HttpError } from "./httpError.ts";
+import { HttpError } from "@shared/http-error.ts";
 import { callAzureFluxKontext } from "./models/azureFluxKontextModel.js";
 import { callFluxKleinAPI } from "./models/fluxKleinModel.ts";
 import {
