+++ gen.pollinations.ai/src/image/models/azureFluxKontextModel.ts
@@ -4,7 +4,7 @@ import type {
     ImageGenerationResult,
 } from "../createAndReturnImages.ts";
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
+import { HttpError } from "@shared/http-error.ts";
 import type { ImageParams } from "../params.ts";
 import { sanitizeString } from "../util.ts";
 import {
