+++ gen.pollinations.ai/src/image/models/openRouterImageModel.ts
@@ -3,7 +3,7 @@ import type { Usage } from "@shared/registry/registry.ts";
 import debug from "debug";
 import type { ImageGenerationResult } from "../createAndReturnImages.ts";
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
+import { HttpError } from "@shared/http-error.ts";
 import type { ImageParams } from "../params.ts";
 import {
     closestAspectRatio,
