+++ gen.pollinations.ai/src/image/utils/imageTransform.ts
@@ -1,4 +1,4 @@
-import { HttpError } from "../httpError.ts";
+import { HttpError } from "@shared/http-error.ts";
 
 type TransformOptions = {
     format?: "image/jpeg" | "image/png" | "image/webp";
