+++ gen.pollinations.ai/src/image/utils/replicateClient.ts
@@ -9,7 +9,7 @@
  */
 
 import { getImageEnv } from "../env.ts";
-import { HttpError } from "../httpError.ts";
+import { HttpError } from "@shared/http-error.ts";
 import { sleep } from "../util.ts";
 
 const API_BASE = "https://api.replicate.com/v1";
