+++ gen.pollinations.ai/src/model3d/handler.ts
@@ -4,7 +4,7 @@ import { buildUsageHeaders } from "@shared/registry/usage-headers.ts";
 import type { Context } from "hono";
 import type { Env } from "@/env.ts";
 import { withModelFallbackResponse } from "../fallback.ts";
-import { HttpError } from "../image/httpError.ts";
+import { HttpError } from "@shared/http-error.ts";
 import { bufferToUint8Array } from "../image/utils/imageDownload.ts";
 import {
     createAndReturnModel3d,
